requirements.txt
beautifulsoup4==4.11.1
python-dotenv==0.20.0
requests==2.27.1
validators==0.18.2
python-slugify==8.0.1